Updated and fixed delete old buckets function, now works as intended

+++ s3deleteOldBuckets/index.js
@@ -15,26 +15,30 @@ exports.handler =  function(event, context) {
             console.log (err, err.stack);
         } else {
             var items = data["Items"];
+            // console.log(items);
             for (item of items){
+                var bucketName = item["bucket_name"]["S"];
                 // variable for tracking if entry and bucket are to be deleted
                 var toDelete = false;
                 // Check if completed
                 if (item["status"] && item["status"]["S"] === "completed"){
+                    console.log(bucketName + " is completed and to be deleted")
                     toDelete = true;
+                // Check if expired
                 } else if (item["expiry_time"] && item["expiry_time"]["N"] < Date.now()){
+                    console.log(bucketName + " is past expiry and to be deleted")
                     toDelete = true;
+                } else {
+                    console.log(bucketName + " is not yet expired so will remain")
                 }
 
-                // Check if expired
 
                 if (toDelete){
-                    var bucketName = item["bucket_name"]["S"];
-                    // delete db entry
-                    deleteDbEntry(bucketName, dynamoDb);
                     // delete bucket
-                    deleteBucket(bucketName);
-                    // Return as nothing left to do
-                    return;
+                    if (deleteBucket(bucketName)){
+                        // delete db entry
+                        deleteDbEntry(bucketName, dynamoDb);
+                    }
                 }
             }
         }
@@ -53,10 +57,11 @@ function deleteDbEntry(bucketName, dynamoDb){
     }
     dynamoDb.deleteItem(deleteParams, function(err,data){
         if (err){
-            console.log("Unsuccessful deletion of db entry");
+            console.log (err, err.stack);
+            console.log("Unsuccessful deletion of db entry for " + bucketName);
             return;
         } else {
-            console.log("Successful deletion of db entry");
+            console.log("Successful deletion of db entry for " + bucketName);
             return;
         }
     })
@@ -70,15 +75,17 @@ function deleteBucket(bucketName){
     }
     s3.listObjectsV2(listParams, function(err, data){
         if (err) {
-            console.log("Unsuccessful action of listing objects in temp bucket");
+            console.log (err, err.stack);
+            console.log("Unsuccessful action of listing objects in temp bucket " + bucketName);
             return;
         } else {
-            console.log("Successful action of listing objects in temp bucket");
+            console.log("Successful action of listing objects in temp bucket " + bucketName);
             // Variable to hold object keys for use in deleteObjects request
             var objectKeys = [];
             var objects = data["Contents"];
             // Go through each object and add its key to the deletion pile
-            for (object in objects){
+            for (object of objects){
+                console.log(object);
                 var key = object["Key"];
                 objectKeys.push({Key:key});
             }
@@ -88,25 +95,28 @@ function deleteBucket(bucketName){
                     Objects: objectKeys
                 }
             }
+            console.log("Delete params: " + JSON.stringify(deleteParams))
             // Delete all objects in the bucket
             s3.deleteObjects(deleteParams, function(err, data){
                 if (err){
-                    console.log("Unsuccessful deletion of bucket objects");
+                    console.log (err, err.stack);
+                    console.log("Unsuccessful deletion of bucket objects for " + bucketName);
                     return;
 
                 } else {
-                    console.log("Successful deletion of bucket objects");
+                    console.log("Successful deletion of bucket objects for " + bucketName);
                     // Delete the bucket
                     var deleteBucketParams = {
                         Bucket: bucketName
                     }
                     s3.deleteBucket(deleteBucketParams, function (err,data){
                         if (err) {
-                            console.log("Unsuccessful deletion of bucket");
+                            console.log (err, err.stack);
+                            console.log("Unsuccessful deletion of bucket " + bucketName);
                             return;
                         } else {
-                            console.log("Successful deletion of bucket objects");
-                            return;
+                            console.log("Successful deletion of bucket objects " + bucketName);
+                            return true;
                         }
                     })
                 }
